Draft initial basegrid setup
Properties and procedures that are common to unstructured grids and
structured grids are defined in the BaseGrid class. This includes
lon, lat, time, time_origin, and mesh properties.

The Grid class is now defined as a type extension of BaseGrid. This
configuration is compatible with the previous commit of Parcels (all
existing tests pass). The _zonal_periodic, _zonal_halo,
_meridional_halo, and _lat_flipped properties are seen as attributes
specific to the structured grid.

The UGrid class is defined as a type extension of the BaseGrid. This
child class adds the `face_node_connectivity` property that relates vertices to the
2-D lateral grid faces. The lon and lat attributes refer to the `node_lon` and
`node_lat` that are defined in the uxarray.Grid structure; these are the
corner node vertices. Note that there are other possible nodes,
including the edge vertices (centers of the edges), and `face_vertices`
that define the element centroids; these are currently not defined.
