+++ parcels/grid.py
@@ -10,6 +10,8 @@ from parcels._typing import Mesh, UpdateStatus, assert_valid_mesh
 from parcels.tools._helpers import deprecated_made_private
 from parcels.tools.converters import TimeConverter
 from parcels.tools.warnings import FieldSetWarning
+from parcels.basegrid import BaseGrid
+from parcels.basegrid import CGrid
 
 __all__ = [
     "GridType",
@@ -21,7 +23,6 @@ __all__ = [
     "CGrid",
     "Grid",
 ]
-
 
 class GridType(IntEnum):
     RectilinearZGrid = 0
@@ -35,100 +36,19 @@ class GridType(IntEnum):
 GridCode = GridType
 
 
-class CGrid(Structure):
-    _fields_ = [("gtype", c_int), ("grid", c_void_p)]
-
-
-class Grid:
+class Grid(BaseGrid):
     """Grid class that defines a (spatial and temporal) grid on which Fields are defined."""
 
-    def __init__(
-        self,
-        lon: npt.NDArray,
-        lat: npt.NDArray,
-        time: npt.NDArray | None,
-        time_origin: TimeConverter | None,
-        mesh: Mesh,
-    ):
-        self._ti = -1
-        self._update_status: UpdateStatus | None = None
-        if not lon.flags["C_CONTIGUOUS"]:
-            lon = np.array(lon, order="C")
-        if not lat.flags["C_CONTIGUOUS"]:
-            lat = np.array(lat, order="C")
-        time = np.zeros(1, dtype=np.float64) if time is None else time
-        if not time.flags["C_CONTIGUOUS"]:
-            time = np.array(time, order="C")
-        if not lon.dtype == np.float32:
-            lon = lon.astype(np.float32)
-        if not lat.dtype == np.float32:
-            lat = lat.astype(np.float32)
-        if not time.dtype == np.float64:
-            assert isinstance(
-                time[0], (np.integer, np.floating, float, int)
-            ), "Time vector must be an array of int or floats"
-            time = time.astype(np.float64)
-
-        self._lon = lon
-        self._lat = lat
-        self.time = time
-        self.time_full = self.time  # needed for deferred_loaded Fields
-        self._time_origin = TimeConverter() if time_origin is None else time_origin
-        assert isinstance(self.time_origin, TimeConverter), "time_origin needs to be a TimeConverter object"
-        assert_valid_mesh(mesh)
-        self._mesh = mesh
-        self._cstruct = None
-        self._cell_edge_sizes: dict[str, npt.NDArray] = {}
+    def __init__(self,*args, **kwargs):
+        super().__init__(*args, **kwargs)
         self._zonal_periodic = False
         self._zonal_halo = 0
         self._meridional_halo = 0
         self._lat_flipped = False
-        self._defer_load = False
-        self._lonlat_minmax = np.array(
-            [np.nanmin(lon), np.nanmax(lon), np.nanmin(lat), np.nanmax(lat)], dtype=np.float32
-        )
-        self.periods = 0
-        self._load_chunk: npt.NDArray = np.array([])
-        self.chunk_info = None
-        self.chunksize = None
-        self._add_last_periodic_data_timestep = False
-        self.depth_field = None
-
-    def __repr__(self):
-        with np.printoptions(threshold=5, suppress=True, linewidth=120, formatter={"float": "{: 0.2f}".format}):
-            return (
-                f"{type(self).__name__}("
-                f"lon={self.lon!r}, lat={self.lat!r}, time={self.time!r}, "
-                f"time_origin={self.time_origin!r}, mesh={self.mesh!r})"
-            )
-
-    @property
-    def lon(self):
-        return self._lon
-
-    @property
-    def lat(self):
-        return self._lat
-
-    @property
-    def depth(self):
-        return self._depth
-
-    @property
-    def mesh(self):
-        return self._mesh
 
     @property
     def meridional_halo(self):
         return self._meridional_halo
-
-    @property
-    def lonlat_minmax(self):
-        return self._lonlat_minmax
-
-    @property
-    def time_origin(self):
-        return self._time_origin
 
     @property
     def zonal_periodic(self):
@@ -137,14 +57,6 @@ class Grid:
     @property
     def zonal_halo(self):
         return self._zonal_halo
-
-    @property
-    def defer_load(self):
-        return self._defer_load
-
-    @property
-    def cell_edge_sizes(self):
-        return self._cell_edge_sizes
 
     @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
@@ -213,12 +125,6 @@ class Grid:
             else:
                 return CurvilinearSGrid(lon, lat, depth, time, time_origin=time_origin, mesh=mesh, **kwargs)
 
-    @property
-    def ctypes_struct(self):
-        # This is unnecessary for the moment, but it could be useful when going will fully unstructured grids
-        self._cgrid = cast(pointer(self._child_ctypes_struct), c_void_p)
-        cstruct = CGrid(self._gtype, self._cgrid.value)
-        return cstruct
 
     @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
@@ -340,88 +246,10 @@ class Grid:
     def computeTimeChunk(self, *args, **kwargs):
         return self._computeTimeChunk(*args, **kwargs)
 
-    def _computeTimeChunk(self, f, time, signdt):
-        nextTime_loc = np.inf if signdt >= 0 else -np.inf
-        periods = self.periods.value if isinstance(self.periods, c_int) else self.periods
-        prev_time_indices = self.time
-        if self._update_status == "not_updated":
-            if self._ti >= 0:
-                if (
-                    time - periods * (self.time_full[-1] - self.time_full[0]) < self.time[0]
-                    or time - periods * (self.time_full[-1] - self.time_full[0]) > self.time[1]
-                ):
-                    self._ti = -1  # reset
-                elif signdt >= 0 and (
-                    time - periods * (self.time_full[-1] - self.time_full[0]) < self.time_full[0]
-                    or time - periods * (self.time_full[-1] - self.time_full[0]) >= self.time_full[-1]
-                ):
-                    self._ti = -1  # reset
-                elif signdt < 0 and (
-                    time - periods * (self.time_full[-1] - self.time_full[0]) <= self.time_full[0]
-                    or time - periods * (self.time_full[-1] - self.time_full[0]) > self.time_full[-1]
-                ):
-                    self._ti = -1  # reset
-                elif (
-                    signdt >= 0
-                    and time - periods * (self.time_full[-1] - self.time_full[0]) >= self.time[1]
-                    and self._ti < len(self.time_full) - 2
-                ):
-                    self._ti += 1
-                    self.time = self.time_full[self._ti : self._ti + 2]
-                    self._update_status = "updated"
-                elif (
-                    signdt < 0
-                    and time - periods * (self.time_full[-1] - self.time_full[0]) <= self.time[0]
-                    and self._ti > 0
-                ):
-                    self._ti -= 1
-                    self.time = self.time_full[self._ti : self._ti + 2]
-                    self._update_status = "updated"
-            if self._ti == -1:
-                self.time = self.time_full
-                self._ti, _ = f._time_index(time)
-                periods = self.periods.value if isinstance(self.periods, c_int) else self.periods
-                if (
-                    signdt == -1
-                    and self._ti == 0
-                    and (time - periods * (self.time_full[-1] - self.time_full[0])) == self.time[0]
-                    and f.time_periodic
-                ):
-                    self._ti = len(self.time) - 1
-                    periods -= 1
-                if signdt == -1 and self._ti > 0 and self.time_full[self._ti] == time:
-                    self._ti -= 1
-                if self._ti >= len(self.time_full) - 1:
-                    self._ti = len(self.time_full) - 2
-
-                self.time = self.time_full[self._ti : self._ti + 2]
-                self.tdim = 2
-                if prev_time_indices is None or len(prev_time_indices) != 2 or len(prev_time_indices) != len(self.time):
-                    self._update_status = "first_updated"
-                elif functools.reduce(
-                    lambda i, j: i and j, map(lambda m, k: m == k, self.time, prev_time_indices), True
-                ) and len(prev_time_indices) == len(self.time):
-                    self._update_status = "not_updated"
-                elif functools.reduce(
-                    lambda i, j: i and j, map(lambda m, k: m == k, self.time[:1], prev_time_indices[:1]), True
-                ) and len(prev_time_indices) == len(self.time):
-                    self._update_status = "updated"
-                else:
-                    self._update_status = "first_updated"
-            if signdt >= 0 and (self._ti < len(self.time_full) - 2 or not f.allow_time_extrapolation):
-                nextTime_loc = self.time[1] + periods * (self.time_full[-1] - self.time_full[0])
-            elif signdt < 0 and (self._ti > 0 or not f.allow_time_extrapolation):
-                nextTime_loc = self.time[0] + periods * (self.time_full[-1] - self.time_full[0])
-        return nextTime_loc
-
     @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def chunk_not_loaded(self):
         return self._chunk_not_loaded
-
-    @property
-    def _chunk_not_loaded(self):
-        return 0
 
     @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
@@ -429,17 +257,9 @@ class Grid:
         return self._chunk_loading_requested
 
     @property
-    def _chunk_loading_requested(self):
-        return 1
-
-    @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def chunk_loaded_touched(self):
         return self._chunk_loaded_touched
-
-    @property
-    def _chunk_loaded_touched(self):
-        return 2
 
     @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
@@ -447,17 +267,9 @@ class Grid:
         return self._chunk_deprecated
 
     @property
-    def _chunk_deprecated(self):
-        return 3
-
-    @property
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def chunk_loaded(self):
         return self._chunk_loaded
-
-    @property
-    def _chunk_loaded(self):
-        return [2, 3]
 
 
 class RectilinearGrid(Grid):
